requirements.txt
hydra-core
omegaconf
torchvision
Pillow
pytest
